menambahkan notifikasi pada pengiriman produk jika ada produk keluar yang perlu dikirim, menambahkan fitur batal pada pengiriman produk jika sopir belum konfirmasi, menghilangkan fitur batal pada pengiriman produk jika sopir sudah konfirmasi, menghilangkan fitur edit dan hapus pada produk keluar jika produk telah dikirim, menambahkan fitur upload bukti foto pada sopir jika sudah selesai mengirim (MIGRATE BANG, kalo mau nyobain fitur yang pengiriman buka 2 browser yang satu login akun backoffice/distribusi yang satu lagi login akun sopir, yang sopir tambahin aja datanya ntar otomatis bisa login pake NIK untuk username dan passwordnya)

+++ app/Http/Controllers/ProdukKeluarController.php
@@ -166,22 +166,31 @@ class ProdukKeluarController extends Controller
     {
         $this->authorize('update', $produkKeluar);
 
-        // join dengan tabel satuan
-        $produkJadi = ProdukJadi::join('satuan', 'produkJadi.kd_satuan', '=', 'satuan.id_satuan')
-            ->select('produkJadi.*', 'satuan.nm_satuan')
-            ->where('kd_produk', $produkKeluar->kd_produk)
-            ->first();
+        // ambil status dari database berdasarkan id
+        $status = ProdukKeluar::where('id_produkKeluar', $produkKeluar->id_produkKeluar)->first()->stts;
 
-        return view(
-            'pages.produkKeluar.edit',
-            ['produkKeluar' => $produkKeluar, 'produkJadi' => $produkJadi],
-            [
-                'tittle' => 'Edit Data',
-                'judul' => 'Edit Penjualan Produk',
-                'menu' => 'Penjualan Produk',
-                'submenu' => 'Edit Data'
-            ]
-        );
+        if ($status == 0) {
+
+            // join dengan tabel satuan
+            $produkJadi = ProdukJadi::join('satuan', 'produkJadi.kd_satuan', '=', 'satuan.id_satuan')
+                ->select('produkJadi.*', 'satuan.nm_satuan')
+                ->where('kd_produk', $produkKeluar->kd_produk)
+                ->first();
+
+            return view(
+                'pages.produkKeluar.edit',
+                ['produkKeluar' => $produkKeluar, 'produkJadi' => $produkJadi],
+                [
+                    'tittle' => 'Edit Data',
+                    'judul' => 'Edit Penjualan Produk',
+                    'menu' => 'Penjualan Produk',
+                    'submenu' => 'Edit Data'
+                ]
+            );
+        } else {
+            Alert::error('Gagal!', 'Data penjualan sudah berada dipengiriman');
+            return redirect('produkKeluar');
+        }
     }
 
     /**
@@ -196,56 +205,65 @@ class ProdukKeluarController extends Controller
 
         $this->authorize('update', $produkKeluar);
 
-        // mengubah nama validasi
-        $messages = [
-            'kd_produk.required' => 'Kode Produk Harus Diisi',
-            'tgl_keluar.required' => 'Tanggal Keluar Harus Diisi',
-            'jumlah.required' => 'Jumlah Harus Diisi',
-            'jumlah.numeric' => 'Jumlah Harus Angka',
-            'ket.required' => 'Keterangan Harus Diisi',
-        ];
+        // ambil status dari database berdasarkan id
+        $status = ProdukKeluar::where('id_produkKeluar', $produkKeluar->id_produkKeluar)->first()->stts;
 
-        $request->validate([
-            'kd_produk' => 'required',
-            'tgl_keluar' => 'required',
-            'jumlah' => 'required|numeric',
-            'ket' => 'required',
-        ], $messages);
+        if ($status == 0) {
+            dd('kesini');
+            // mengubah nama validasi
+            $messages = [
+                'kd_produk.required' => 'Kode Produk Harus Diisi',
+                'tgl_keluar.required' => 'Tanggal Keluar Harus Diisi',
+                'jumlah.required' => 'Jumlah Harus Diisi',
+                'jumlah.numeric' => 'Jumlah Harus Angka',
+                'ket.required' => 'Keterangan Harus Diisi',
+            ];
 
-        $nip = auth()->user()->nip;
+            $request->validate([
+                'kd_produk' => 'required',
+                'tgl_keluar' => 'required',
+                'jumlah' => 'required|numeric',
+                'ket' => 'required',
+            ], $messages);
 
-        // mengembalikan stok produk
-        $stok = ProdukJadi::where('kd_produk', $produkKeluar->kd_produk)->first();
-        $stok->stok = $stok->stok + $produkKeluar->jumlah;
-        $stok->save();
+            $nip = auth()->user()->nip;
 
-        // update stok produk
-        $stok = ProdukJadi::where('kd_produk', $request->kd_produk)->first();
-        $stok->stok = $stok->stok - $request->jumlah;
-        $stok->save();
+            // mengembalikan stok produk
+            $stok = ProdukJadi::where('kd_produk', $produkKeluar->kd_produk)->first();
+            $stok->stok = $stok->stok + $produkKeluar->jumlah;
+            $stok->save();
 
-        // ubah format tgl_keluar dari varchar ke date
-        $tgl_keluar = date('Y-m-d', strtotime($request->tgl_keluar));
+            // update stok produk
+            $stok = ProdukJadi::where('kd_produk', $request->kd_produk)->first();
+            $stok->stok = $stok->stok - $request->jumlah;
+            $stok->save();
 
-        $harga_jual = ProdukJadi::where('kd_produk', $request->kd_produk)->first()->harga_jual;
+            // ubah format tgl_keluar dari varchar ke date
+            $tgl_keluar = date('Y-m-d', strtotime($request->tgl_keluar));
 
-        $total = $harga_jual * $request->jumlah;
+            $harga_jual = ProdukJadi::where('kd_produk', $request->kd_produk)->first()->harga_jual;
 
-        $stts = 0;
+            $total = $harga_jual * $request->jumlah;
 
-        $produkKeluar->update([
-            'kd_produk' => $request->kd_produk,
-            'nip_karyawan' => $nip,
-            'tgl_keluar' => $tgl_keluar,
-            'harga_jual' => $harga_jual,
-            'jumlah' => $request->jumlah,
-            'total' => $total,
-            'ket' => $request->ket,
-            'stts' => $stts,
-        ]);
+            $stts = 0;
 
-        Alert::success('Data Penjualan Produk', 'Berhasil Diubah!');
-        return redirect('produkKeluar');
+            $produkKeluar->update([
+                'kd_produk' => $request->kd_produk,
+                'nip_karyawan' => $nip,
+                'tgl_keluar' => $tgl_keluar,
+                'harga_jual' => $harga_jual,
+                'jumlah' => $request->jumlah,
+                'total' => $total,
+                'ket' => $request->ket,
+                'stts' => $stts,
+            ]);
+
+            Alert::success('Data Penjualan Produk', 'Berhasil Diubah!');
+            return redirect('produkKeluar');
+        } else {
+            Alert::error('Gagal!', 'Data penjualan sudah berada dipengiriman');
+            return redirect('produkKeluar');
+        }
     }
 
     /**
@@ -258,13 +276,21 @@ class ProdukKeluarController extends Controller
     {
         $this->authorize('delete', $produkKeluar);
 
-        // update stok produk
-        $stok = ProdukJadi::where('kd_produk', $produkKeluar->kd_produk)->first();
-        $stok->stok = $stok->stok + $produkKeluar->jumlah;
-        $stok->save();
+        // ambil status dari database berdasarkan id
+        $status = ProdukKeluar::where('id_produkKeluar', $produkKeluar->id_produkKeluar)->first()->stts;
 
-        $produkKeluar->delete();
-        Alert::success('Data Penjualan Produk', 'Berhasil Dihapus!');
-        return redirect('produkKeluar');
+        if ($status == 0) {
+            // update stok produk
+            $stok = ProdukJadi::where('kd_produk', $produkKeluar->kd_produk)->first();
+            $stok->stok = $stok->stok + $produkKeluar->jumlah;
+            $stok->save();
+
+            $produkKeluar->delete();
+            Alert::success('Data Penjualan Produk', 'Berhasil Dihapus!');
+            return redirect('produkKeluar');
+        } else {
+            Alert::error('Gagal!', 'Data penjualan sudah berada di Pengiriman');
+            return redirect('produkKeluar');
+        }
     }
 }
